move utils back to original locations
